@@ -23,13 +23,15 @@
   },
   "devDependencies": {
     "@tsconfig/node18": "^1.0.1",
+    "@types/cors": "^2.8.5",
     "@types/express": "^4.17.13",
     "@types/express-serve-static-core": "^4.17.13",
     "@types/node": "^18.11.9",
+    "@types/uuid": "^9.0.0",
     "nodemon": "^2.0.19",
     "prisma": "4.12.0",
     "standard": "^17.0.0",
-    "typescript": "^4.8.4"
+    "typescript": "^5.1.0"
   },
   "engineStrict": true,
   "engines": {
