@@ -13,27 +13,28 @@
   },
   "dependencies": {
     "@prisma/client": "4.12.0",
-    "@tanstack/react-query": "^4.13.0",
-    "axios": "^0.27.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@tanstack/react-query": "^4.29.0",
+    "axios": "^1.4.0",
+    "mitt": "3.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^5.3.3",
     "superjson": "^1.12.2"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
-    "@tsconfig/vite-react": "^1.0.1",
-    "@types/react": "^17.0.53",
-    "@types/react-dom": "^17.0.19",
+    "@testing-library/react": "^14.0.0",
+    "@tsconfig/vite-react": "^2.0.0",
+    "@types/react": "^18.0.37",
+    "@types/react-dom": "^18.0.11",
     "@types/react-router-dom": "^5.3.3",
     "@vitejs/plugin-react-swc": "^3.0.0",
     "@vitest/ui": "^0.29.3",
     "dotenv": "^16.0.3",
     "jsdom": "^21.1.1",
     "msw": "^1.1.0",
-    "typescript": "^4.9.3",
-    "vite": "^4.1.0",
+    "typescript": "^5.1.0",
+    "vite": "^4.3.9",
     "vitest": "^0.29.3"
   },
   "engineStrict": true,
