@@ -1,6 +1,6 @@
 {{={= =}=}}
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import { QueryClientProvider } from '@tanstack/react-query'
 
 import router from './router'
@@ -12,6 +12,10 @@ import {
 {=# setupFn.isDefined =}
 {=& setupFn.importStatement =}
 {=/ setupFn.isDefined =}
+
+{=# areWebSocketsUsed =}
+import { WebSocketProvider } from './webSocket/WebSocketProvider'
+{=/ areWebSocketsUsed =}
 
 startApp()
 
@@ -26,10 +30,18 @@ async function startApp() {
 
 async function render() {
   const queryClient = await queryClientInitialized
-  ReactDOM.render(
-    <QueryClientProvider client={queryClient}>
-      {router}
-    </QueryClientProvider>,
-    document.getElementById('root')
+  ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
+    <React.StrictMode>
+      <QueryClientProvider client={queryClient}>
+        {=# areWebSocketsUsed =}
+        <WebSocketProvider>
+          {router}
+        </WebSocketProvider>
+        {=/ areWebSocketsUsed =}
+        {=^ areWebSocketsUsed =}
+        {router}
+        {=/ areWebSocketsUsed =}
+      </QueryClientProvider>
+    </React.StrictMode>
   )
 }
