@@ -1,7 +1,8 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import { Link } from 'react-router-dom'
 import { User } from '@wasp/auth/types'
 import api from '@wasp/api'
+import { useSocket, useSocketListener, ServerToClientPayload } from '@wasp/webSocket'
 
 async function fetchCustomRoute() {
   const res = await api.get('/foo/bar')
@@ -13,9 +14,31 @@ export const ProfilePage = ({
 }: {
   user: User
 }) => {
+  const [messages, setMessages] = useState<ServerToClientPayload<'chatMessage'>[]>([])
+  const { socket, isConnected } = useSocket()
+  const inputRef = useRef<HTMLInputElement>(null)
+
   useEffect(() => {
     fetchCustomRoute()
   }, [])
+
+  useSocketListener('chatMessage', (msg) => setMessages((priorMessages) => [msg, ...priorMessages]))
+
+  function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
+    e.preventDefault()
+
+    if (inputRef.current !== null) {
+      socket.emit('chatMessage', inputRef.current.value)
+      inputRef.current.value = ''
+    }
+  }
+
+  const messageList = messages.map((msg) => (
+    <li key={msg.id}>
+      <em>{msg.username}</em>: {msg.text}
+    </li>
+  ))
+  const connectionIcon = isConnected ? '🟢' : '🔴'
 
   return (
     <>
@@ -26,6 +49,22 @@ export const ProfilePage = ({
       </div>
       <br />
       <Link to="/">Go to dashboard</Link>
+      <div>
+        <form onSubmit={handleSubmit}>
+          <div className="flex space-x-4 place-items-center">
+            <div>{connectionIcon}</div>
+            <div>
+              <input type="text" ref={inputRef} />
+            </div>
+            <div>
+              <button className="btn btn-primary" type="submit">
+                Submit
+              </button>
+            </div>
+          </div>
+        </form>
+        <ul>{messageList}</ul>
+      </div>
     </>
   )
 }
